requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==1.9
Flask==0.12.2
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
pipenv==11.9.0
pycparser==2.18
PyJWT==1.6.1
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
